Split packages, replaced shift with slice
 Changes to be committed:
	modified:   fabric-rendering-v1/src/main/java/net/fabricmc/fabric/api/client/render/v1/BlockEntityRendererRegistry.java
	modified:   fabric-rendering-v1/src/main/java/net/fabricmc/fabric/api/client/render/v1/EntityRendererRegistry.java
	new file:   fabric-rendering-v1/src/main/java/net/fabricmc/fabric/impl/client/render/BlockEntityRendererRegistryImpl.java
	new file:   fabric-rendering-v1/src/main/java/net/fabricmc/fabric/impl/client/render/EntityRendererRegistryImpl.java
	modified:   fabric-rendering-v1/src/main/java/net/fabricmc/fabric/mixin/client/rendering/MixinBlockEntityRenderDispatcher.java
	modified:   fabric-rendering-v1/src/main/java/net/fabricmc/fabric/mixin/client/rendering/MixinEntityRenderManager.java
