+++ fabric-rendering-v1/src/main/java/net/fabricmc/fabric/api/client/render/v1/BlockEntityRendererRegistry.java
@@ -16,49 +16,16 @@
 
 package net.fabricmc.fabric.api.client.render.v1;
 
-import java.util.HashMap;
-import java.util.Map;
-
 import net.minecraft.block.entity.BlockEntity;
-import net.minecraft.client.render.block.entity.BlockEntityRenderDispatcher;
 import net.minecraft.client.render.block.entity.BlockEntityRenderer;
+
+import net.fabricmc.fabric.impl.client.render.BlockEntityRendererRegistryImpl;
 
 /**
  * Helper class for registering BlockEntityRenderers.
  */
-public class BlockEntityRendererRegistry {
-	public static final BlockEntityRendererRegistry INSTANCE = new BlockEntityRendererRegistry();
-	private Map<Class<? extends BlockEntity>, BlockEntityRenderer<? extends BlockEntity>> renderers = null;
-	private Map<Class<? extends BlockEntity>, BlockEntityRenderer<? extends BlockEntity>> renderersTmp = new HashMap<>();
+public interface BlockEntityRendererRegistry {
+	BlockEntityRendererRegistry INSTANCE = new BlockEntityRendererRegistryImpl();
 
-	private BlockEntityRendererRegistry() {
-	}
-
-	public void initialize(BlockEntityRenderDispatcher instance, Map<Class<? extends BlockEntity>, BlockEntityRenderer<? extends BlockEntity>> map) {
-		if (renderers != null && renderers != map) {
-			throw new RuntimeException("Tried to set renderers twice!");
-		}
-
-		if (renderers == map) {
-			return;
-		}
-
-		renderers = map;
-
-		for (BlockEntityRenderer<? extends BlockEntity> renderer : renderersTmp.values()) {
-			renderer.method_3751(instance);
-		}
-
-		renderers.putAll(renderersTmp);
-		renderersTmp = null;
-	}
-
-	public void register(Class<? extends BlockEntity> blockEntityClass, BlockEntityRenderer<? extends BlockEntity> blockEntityRenderer) {
-		if (renderers != null) {
-			renderers.put(blockEntityClass, blockEntityRenderer);
-			blockEntityRenderer.method_3751(BlockEntityRenderDispatcher.INSTANCE);
-		} else {
-			renderersTmp.put(blockEntityClass, blockEntityRenderer);
-		}
-	}
+	void register(Class<? extends BlockEntity> blockEntityClass, BlockEntityRenderer<? extends BlockEntity> blockEntityRenderer);
 }
