+++ fabric-rendering-v1/src/main/java/net/fabricmc/fabric/api/client/render/v1/EntityRendererRegistry.java
@@ -16,74 +16,17 @@
 
 package net.fabricmc.fabric.api.client.render.v1;
 
-import java.util.HashMap;
-import java.util.Map;
-import java.util.WeakHashMap;
-
 import net.minecraft.client.render.entity.EntityRenderDispatcher;
 import net.minecraft.client.render.entity.EntityRenderer;
-import net.minecraft.client.render.item.ItemRenderer;
-import net.minecraft.client.texture.TextureManager;
 import net.minecraft.entity.Entity;
+
+import net.fabricmc.fabric.impl.client.render.EntityRendererRegistryImpl;
 
 /**
  * Helper class for registering EntityRenderers.
  */
-public class EntityRendererRegistry {
-	@FunctionalInterface
-	public interface Factory {
-		EntityRenderer<? extends Entity> create(EntityRenderDispatcher manager, Context context);
-	}
+public interface EntityRendererRegistry {
+	EntityRendererRegistry INSTANCE = new EntityRendererRegistryImpl();
 
-	public static final class Context {
-		private final TextureManager textureManager;
-		private final ItemRenderer itemRenderer;
-		private final Map<Class<? extends Entity>, EntityRenderer<? extends Entity>> rendererMap;
-
-		private Context(TextureManager textureManager, ItemRenderer itemRenderer, Map<Class<? extends Entity>, EntityRenderer<? extends Entity>> rendererMap) {
-			this.textureManager = textureManager;
-			this.itemRenderer = itemRenderer;
-			this.rendererMap = rendererMap;
-		}
-
-		public TextureManager getTextureManager() {
-			return textureManager;
-		}
-
-		public ItemRenderer getItemRenderer() {
-			return itemRenderer;
-		}
-	}
-
-	public static final EntityRendererRegistry INSTANCE = new EntityRendererRegistry();
-	private final Map<EntityRenderDispatcher, Context> renderManagerMap = new WeakHashMap<>();
-	private final Map<Class<? extends Entity>, Factory> renderSupplierMap = new HashMap<>();
-
-	private EntityRendererRegistry() {
-	}
-
-	public void initialize(EntityRenderDispatcher manager, TextureManager textureManager, ItemRenderer itemRenderer, Map<Class<? extends Entity>, EntityRenderer<? extends Entity>> map) {
-		synchronized (renderSupplierMap) {
-			if (renderManagerMap.containsKey(manager)) {
-				return;
-			}
-
-			Context context = new Context(textureManager, itemRenderer, map);
-			renderManagerMap.put(manager, context);
-
-			for (Class<? extends Entity> c : renderSupplierMap.keySet()) {
-				map.put(c, renderSupplierMap.get(c).create(manager, context));
-			}
-		}
-	}
-
-	public void register(Class<? extends Entity> entityClass, Factory factory) {
-		synchronized (renderSupplierMap) {
-			renderSupplierMap.put(entityClass, factory);
-
-			for (EntityRenderDispatcher manager : renderManagerMap.keySet()) {
-				renderManagerMap.get(manager).rendererMap.put(entityClass, factory.create(manager, renderManagerMap.get(manager)));
-			}
-		}
-	}
+	void register(Class<? extends Entity> entityClass, EntityRendererRegistryImpl.Factory factory);
 }
